leveraging agents to work faster, but created initial scaffolding and abstracting of basic functions out to have CRUD functions for each of the tables, with additional analysis on the tables like the contador, etc.

+++ supabase/schemas/06_packing_lists.sql
@@ -1,7 +1,7 @@
 create table public.packing_lists (
   id bigint generated always as identity primary key,
   num_pl integer not null,
-  customer_id integer references public.customers (id),
+  customer_id bigint references public.customers (id),
   customer text,
   date date,
   ship_date date,
